requirements.txt
flask
flask_session
qrcode
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
email-validator
Pillow
geopy
werkzeug
sqlite3
uuid
math
smtplib
functools
secrets
string
base64
collections
request
jsonify
send_from_directory
current_app
send_file
time
datetime
io
io.BytesIO
Image
ImageDraw
werkzeug
werkzeug.user_agent
werkzeug.security
werkzeug.utils
email.message.EmailMessage
google.oauth2.credentials.Credentials
google.auth.transport.requests.Request
googleapiclient.discovery.build
googleapiclient.errors.HttpError
